pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-actor"
version = "0.1.0"
description = "Thin Pipecat wrapper for Deepgram Flux → Gemini → Deepgram Aura voice pipeline"
authors = [{name = "Jan Zuiderveld"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aenum==3.1.16",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "dataclasses-json==0.6.7",
    "deepgram-sdk==4.7.0",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "docstring_parser==0.17.0",
    "frozenlist==1.8.0",
    "future==1.0.0",
    "google-api-core==2.26.0",
    "google-auth==2.41.1",
    "google-cloud-speech==2.34.0",
    "google-cloud-texttospeech==2.32.0",
    "google-genai==1.46.0",
    "googleapis-common-protos==1.71.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jiter==0.11.1",
    "joblib==1.5.2",
    "llvmlite==0.44.0",
    "loguru==0.7.3",
    "Markdown==3.9",
    "markdown-it-py==4.0.0",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "multidict==6.7.0",
    "mypy_extensions==1.1.0",
    "nltk==3.9.2",
    "numba==0.61.2",
    "numpy==2.2.6",
    "openai==2.6.0",
    "packaging==25.0",
    "pillow==11.3.0",
    "pipecat-ai[deepgram,google,local]==0.0.91",
    "pipecat-thin==0.1.0",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "PyAudio==0.2.14",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic_core==2.41.4",
    "Pygments==2.19.2",
    "pyloudnorm==0.1.1",
    "pyserial==3.5",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "python-dotenv==1.1.1",
    "regex==2025.10.23",
    "requests==2.32.5",
    "resampy==0.4.3",
    "rich==14.2.0",
    "rsa==4.9.1",
    "scipy==1.16.2",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sounddevice==0.5.3",
    "soxr==0.5.0.post1",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typer==0.20.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.5.0",
    "wait_for2==0.4.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "yarl==1.22.0",
]

[project.optional-dependencies]
test = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
